Isolate CucumberRailsHelper in its own file and tidy up a couple of support code lines
